Break up reservation.py to improve separation of concerns.
